USART DMA bug
When DMA length was greater than 255 bytes, TM_USART_DMA_Sending
function did not work properly, since it returned 0 when NDTR has a
value which is multiply of 256.

+++ 00-STM32F429_LIBRARIES/tm_stm32f4_usart_dma.h
@@ -270,7 +270,7 @@ uint8_t TM_USART_DMA_Send(USART_TypeDef* USARTx, uint8_t* DataArray, uint16_t co
  *            - 0: USART does not sending anymore
  *            - > 0: USART DMA is still sending data
  */
-uint8_t TM_USART_DMA_Sending(USART_TypeDef* USARTx);
+uint16_t TM_USART_DMA_Sending(USART_TypeDef* USARTx);
 
 /**
  * @}
